requirements.txt
 
python-dotenv
transformers
openai
langchain-community
torch
streamlit
tiktoken